* Added MethodSymbol StackLocals counting * Added parameters counting to DeclarationCollector * Added variable counting to DeclarationCollector * Fixed linking, added SHARD_API to new classes

+++ ShardScript/include/shard/syntax/symbols/VariableSymbol.h
@@ -18,7 +18,7 @@ namespace shard
 		const TypeSymbol* Type = nullptr;
         shard::ExpressionSyntax* Declaration = nullptr;
 
-		int SlotIndex = 0;
+		uint16_t SlotIndex = 0;
         bool IsConst = false;
 
 		inline VariableSymbol(std::wstring name, TypeSymbol* type) : SyntaxSymbol(name, SyntaxKind::VariableStatement), Type(type)
